+++ HISTORY.md
@@ -1,6 +1,10 @@
 History
 =======
 
+## Current
+* Change `host` short flag to `-H` (don't conflict with help `-h`).
+* Add Handlebars template support.
+
 ## v0.0.4
 * Tweak PR URLs to be actual PR web pages for normal / Enterprise GitHub.